backend/requirements.txt
django>=6.0,<7.0
djangorestframework>=3.17,<4.0
django-cors-headers>=4.9,<5.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
gunicorn>=23.0,<24.0
